#42 #51 Program now catches OutOfMemoryError
Initially the plan was to impose limits on the size and number of variables in the program, as well as the RAM used, but catching OutOfMemoryError pretty much achieves the same goal of ensuring that the program will not error out if all the memory is used up. I'm happy to just close these issue and call them done.